backend/requirements.txt
fastapi
uvicorn
transformers>=4.30.0
datasets
peft>=0.4.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu
PyPDF2>=3.0.0
python-dotenv
PyYAML
pydantic-settings
